[syncModule]: moved from typedDict to pydantic models, splitdata function working perfectly
